% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gbs2plink.R
\name{convert_gbs2plink}
\alias{convert_gbs2plink}
\alias{run_iupac}
\title{\code{Change gbs hmp to plink format}}
\usage{
convert_gbs2plink(gbsfile = "g2f_2014_zeagbsv27v5hmp_nodash.txt",
  gbs = NULL, iupac = TRUE, mode = 1,
  outfile = "~/dbcenter/seeds_data/chr10_filetered_unimputed.bed")

run_iupac(subg)
}
\arguments{
\item{gbsfile}{An input gbs in hmp format. Note without # lines. [chr, ="g2f_2014_zeagbsv27v5hmp_nodash.txt"]}

\item{gbs}{If gbsfile is NULL load this gbs object. [obj, data.table=NULL]}

\item{iupac}{Change the IUPAC Code. [logical, =TRUE]}

\item{mode}{GBS code mode. mode=1, A, T, C, G; mode=2, AA, TT, CC, GG. [num, =1]}

\item{outfile}{Output file name. [chr, ="~/dbcenter/seeds_data/chr10_filetered_unimputed.bed"]}

\item{jmph}{plink --tped test.tped --tfam test.tfam --make-bed --missing-genotype N --allow-no-sex --make-founders --out ../Data/sim}
}
\description{
load gbs file, change to tfile format for PLINK, filtering multiple alleles and
then change to "A A" coding and change IUPAC Ambiguity Codes
}
