% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack.R
\name{plot_stack}
\alias{plot_stack}
\title{\code{Stacking plot with 10 maize chromosomes}}
\usage{
plot_stack(df, df2, chrlen = "data/ZmB73_RefGen_v2.length", centcsv = NULL,
  plotcol = "Dm", dorescale = FALSE, cols = list("red", rep(c("slateblue",
  "cyan4"), 5), "#ff7373"), ...)
}
\arguments{
\item{df}{An input data.frame, with columns of chr and pos.}

\item{chrlen}{The chr length file, tab delimited.}

\item{centcsv}{Centeromere file, csv.}

\item{plotcol}{The column name to plot.}

\item{dorescale}{Whether to do rescale or not. Default=FALSE.}

\item{cols}{A list of cols, cols[[1]], cent box; cols[[2]], chr; cols[[3]], highlight.}

\item{...}{Other parameters will pass to points().}
}
\value{
return a plot.
}
\description{
\code{Stacking plot with 10 maize chromosomes}
}
\examples{

}
