% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aspera.R
\name{run_aspera}
\alias{run_aspera}
\alias{set_aspera}
\title{\code{Run Aspera Connect to download from SRA.}}
\usage{
run_aspera(sradf, maxspeed = "100m", outdir = ".", cmdno = 1,
  email = NULL, runinfo)

set_aspera(sradf, srow, erow, j, maxspeed, outdir)
}
\arguments{
\item{sradf}{An input data.frame for SRA ids. Must contains column: SRR. [df["SRR"]]}

\item{maxspeed}{The max speed for downloading. [chr, =200m]}

\item{outdir}{The output directory. [chr, ="."]}

\item{cmdno}{Number of commands to excute in each array job. [num, =1]}

\item{email}{Your email address that farm will email to once the job was done/failed. [chr, =NULL]}

\item{runinfo}{[vector, = c(FALSE, "bigmemh", 5, "5G", "16:00:00")]}

\item{srow}{Start row of the df. [num, 1]}

\item{erow}{End row of the df. [num, 10]}
}
\value{
return a batch of shell scripts.
}
\description{
Downloading SRA using 'ascp' utility or Aspera Connect.
How to download and install 'ascp':
\url{http://downloads.asperasoft.com/en/downloads/8?list}
}
\details{
see more detail about SRA with Aspera downloading:
\url{http://www.ncbi.nlm.nih.gov/books/NBK158899/#SRA_download.downloading_sra_data_using}
}
\examples{
sra <- data.frame(SRR=c("ERR877647", "ERR877648"),
SRX=c( "ERX957210", "ERX957211"),
pid=c( "1_Base1_Bbreve-sc-2188486", "P1_ECvsrS_1-sc-2201977"))
run_aspera(sra, maxspeed="200m", outdir=".", arrayjobs="1-2", jobid="aspera", email=NULL)

}
