% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bismark.R
\name{run_bismark}
\alias{run_bismark}
\title{\code{Run array Bismark job on farm}}
\usage{
run_bismark(inputdf, genome = "/home/jolyang/dbcenter/AGP/AGPv2",
  outdir = "/group/jrigrp4/BS_teo20/WGBS/BSM", N = 1, align = TRUE,
  email = NULL, runinfo = c(FALSE, "batch", 1, "10G", "8:00:00"))
}
\arguments{
\item{inputdf}{An input data.frame object. Must contains fq1, fq2 and outbase, bam (optional). [df, ["fq1", "fq2", "outbase", ("bam", "genome")]]}

\item{genome}{The folder of genome prepared by bismark.
If genome is NULL, then genome will read from inputdf column: 'genome'. [chr, ="/home/jolyang/dbcenter/AGP/AGPv2"]}

\item{outdir}{Folder for output. [chr, ="/group/jrigrp4/BS_teo20/WGBS/BSM"]}

\item{N}{Number of mismatches. Sets the number of mismatches to allowed in a seed alignment during multiseed alignment.
Can be set to 0 or 1. Setting this higher makes alignment slower (often much slower)
but increases sensitivity. Default: 0. This option is only available for Bowtie 2 (for Bowtie 1 see -n). [num, =1]}

\item{align}{Whether to conduct alignment, default=TRUE. [logical, =TRUE]}

\item{email}{Your email address that farm will email to once the job was done/failed.}
}
\value{
return a batch of shell scripts.
}
\description{
required modules to load:
bowtie2/2.2.5
bismark/0.14.3
}
\details{
Allow one mismatch 'n 1'

see more detail about Bismark:
\url{http://www.bioinformatics.babraham.ac.uk/projects/bismark/Bismark_User_Guide.pdf}

(I) Running bismark_genome_preparation
module load bismark/0.19
module load bowtie2/2.2
bismark_genome_preparation --bowtie2 /home/jolyang/dbcenter/AGP/AGPv2/

(II) Running bismark

#uses 0-based genomic start and 1-based end coordinates.
bismark_methylation_extractor -s --bedGraph --counts --buffer_size 10G --CX
--cytosine_report --genome_folder /home/jolyang/dbcenter/AGP/AGPv2 test_pe.bam
#<chromosome> <position> <strand> <count methylated> <count unmethylated> <C-context> <trinucleotide context>
}
\examples{
input_df <- data.frame(fq1=c("f_1.fq", "t_1.fq"), fq2=c("f_1.fq", "t_2.fq"), out=c("t1", "t2"))
runa_bismark(input_df, genome="/home/jolyang/dbcenter/AGP/AGPv2",
cpu=4, outdir="/group/jrigrp4/BS_teo20/WGBS/BSM", arrayjobs="1-5", jobid="bs1-5",
email=NULL)
}
