% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bwamem.R
\name{run_bwamem}
\alias{run_bwamem}
\title{\code{Run BWA-mem job on farm}}
\usage{
run_bwamem(fq, kitpath = "/home/jolyang/bin/fermikit", s = "3g", t = 16,
  l = 100, arrayjobs = "1-2", jobid = "fermi", email = NULL)
}
\arguments{
\item{fq}{An input data.frame for fastq files. Must contains fq1, fq2 and out.}

\item{kitpath}{The absolute or relative path of the fermi.kit directory that can invoke the pipeline.}

\item{s}{Approximate genome size, default=3g.}

\item{t}{Number of threads, default=16.}

\item{l}{Primary read length, default=100.}

\item{arrayjobs}{A character specify the number of array you try to run, i.e. 1-100.}

\item{jobid}{The job name show up in your sq NAME column.}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{genome}{The full path of genome with bwa indexed reference fasta file.}
}
\value{
return a batch of shell scripts.
}
\description{
Fermikit is a de novo assembly based variant calling pipeline for Illumina short reads
}
\details{
see more detail about fermikit by Li, Heng:
\url{https://github.com/lh3/fermikit}
}
\examples{
fq <- data.frame(fq1=c("f_1.fq", "t_1.fq"), fq2=c("f_1.fq", "t_2.fq"), out=c("t1", "t2"))
run_fermikit(fq, kitpath="/home/jolyang/bin/fermikit/",
genome="/home/jolyang/dbcenter/AGP/AGPv2", s='3g', t=16, l=100, arrayjobs="1-2",
jobid="fermi", email=NULL)

}
