% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastq_qc.R
\name{run_fastq_qc}
\alias{run_fastq_qc}
\title{\code{Run fastq QC.}}
\usage{
run_fastq_qc(df, method = "seqtk", q = 20, email = NULL,
  runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{df}{Data.frame with fq and out columns.
fq is the path of your fastq files and out is the path of your output file.}

\item{method}{Method for QC, "seqtk" or "FastQC" only.}

\item{q}{Default=20. Note:use -q0 to get the distribution of all quality values}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}

\item{genomesize}{Maize genome size, default=2500000000.}
}
\value{
return A batch of shell scripts.
}
\description{
Run this 'seqtk fqchk -q 20' for all your fastq files.
}
\details{
dependency:
seqtk version 1.0-r82-dirty
}
\examples{
fqs <- c("$HOME/dbcenter/Ecoli/fastq/SRR2921970.sra_1.fastq.gz",
         "$HOME/dbcenter/Ecoli/fastq/SRR2921970.sra_2.fastq.gz")
df <- data.frame(fq=fqs, out=fqs)
df$out <- gsub("sra_.*", "qc", df$out)
run_fastq_qc(df, email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

