% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GenSel4.R
\name{run_GenSel4}
\alias{run_GenSel4}
\title{\code{Run GenSel on slurm-based HPC}}
\usage{
run_GenSel4(inputdf, cv = FALSE, bayesType = "BayesC",
  inpdir = "largedata/", cmdno = 1,
  shid = "slurm-script/run_gensel_array.sh", remove = TRUE, email = NULL,
  runinfo = c(FALSE, "batch", "30", 3, "8:00:00"))
}
\arguments{
\item{inputdf}{An input data.frame for this function, including all the necessary pheno, geno file pathes,
and other parameters. [data.frame, see exaples]}

\item{cv}{Cross-validation experiment [TRUE/FALSE, default=FALSE].
If TRUE, inputdf must contain `trainpheno` and `testpheno`; if FALSE, only need `pheno`.}

\item{bayesType}{Analysis types for Bayes Methods. [chr, default=BayesC; BayesA, BayesB, BayesC, BayesCPi, RBR]}

\item{inpdir}{Dir for the GS .inp file. [chr, default="largedata/"].}

\item{cmdno}{Number of commands per CPU, i.e. number of rows per inputdf. [num, default=1].}

\item{shid}{The shell ID for this job. [chr, default is "slurm-script/run_gensel_array.sh"].}

\item{remove}{Whether to remove the intermediate GS files or not. [TRUE/FALSE, default is TRUE].}

\item{email}{Your email address that farm will email to once the jobs were done/failed.
[chr, default is NULL]}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "batch", "30G", 1, "8:00:00"): 1) run or not, default=FALSE
2) -p partition name, default=batch, 3) --ntasks=3, 4) --mem=30G, and 5) --time=8:00:00.}
}
\value{
return a batch of shell scripts.
}
\description{
GenSel4R
}
\examples{

inputdf <- data.frame(pi=0.995,
   geno="/Users/yangjl/Documents/GWAS2_KRN/SNP/merged/geno_chr.newbin",
   trainpheno="/Users/yangjl/Documents/Heterosis_GWAS/pheno2011/reports/cd_GenSel_fullset.txt",
   testpheno="/Users/yangjl/Documents/Heterosis_GWAS/pheno2011/reports/cd_GenSel_fullset.txt",
   chainLength=1000, burnin=100, varGenotypic=1.4, varResidual=2,
   out="test_out")

run_GenSel4(inputdf, cv=FALSE, inpdir="slurm-script/", cmdno=1,
            shid = "slurm-script/run_gensel_array.sh", remove=TRUE,
            email=NULL, runinfo = c(FALSE, "batch", 3, "30", "8:00:00") )

}
