% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto.R
\name{run_kallisto}
\alias{run_kallisto}
\title{\code{Run kallisto quant.}}
\usage{
run_kallisto(df, single = TRUE, l = 200, s = 20,
  idx = "~/dbcenter/AGP/AGPv2/ZmB73_5a_WGS_exons.kallisto.idx",
  email = NULL, jobid = "run_kquant", runinfo = c(TRUE, "jclarke",
  "1", "8G", "8:00:00"))
}
\arguments{
\item{df}{Input files, [data.frame, col: outdir, fq (SE) or fq1 and fq2 (PE)]}

\item{single}{Single (or PE) end fastq [TRUE(FALSE)].}

\item{l}{Estimated average fragment length [number, default=200].}

\item{s}{Estimated standard deviation of fragment length [number, default=20].}

\item{idx}{Full path of Kallisto Index file, [chr, "/lustre/work/jyanglab/jyang21/dbcenter/AGP/AGPv4/Zea_mays.AGPv4.cdna.all.kallisto.idx"]}

\item{run}{Parameters control the array job partition.
A vector of c(TRUE, "bigmemh", "1", "8:00:00"): 1) run or not, 2) -p partition name, 3) --ntasks and 4) --times.}
}
\value{
return a batch of shell scripts to run.
}
\description{
kallisto quant
}
\details{
see more detail about fastq-dump with Aspera downloading:
\url{https://pachterlab.github.io/kallisto/manual}
}
\examples{
### build kallisto idex
 kallisto index -i ZmB73_5a_WGS_exons.kallisto.idx ZmB73_5a_WGS_exons.fasta.gz
 kallisto quant -i ~/dbcenter/AGP/AGPv2/ZmB73_5a_WGS_exons.kallisto.idx --plaintext -o .
--single -l 200 -s 20 SRR957415.sra.fastq SRR957416.sra.fastq > test.txt

}
