% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_VCAP.R
\name{run_ldak}
\alias{run_ldak}
\title{\code{Set up aspera code in a job shell}}
\usage{
run_ldak(df, email = NULL, runinfo = c(TRUE, "jclarke", "1", "8G",
  "8:00:00"))
}
\arguments{
\item{df}{Start row of the df. [data.frame, cols:output, klist, pheno]}
}
\description{
\code{Set up aspera code in a job shell}
}
