% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Rcode.R
\name{run_Rcode}
\alias{run_Rcode}
\title{\code{Run array job of R codes}}
\usage{
run_Rcode(inputdf, outdir, cmdno = 1, rcodes = "lib/C_format.R",
  rversion = "3.5", arrayshid = "slurm-script/run_bcf_query_array.sh",
  email = NULL, runinfo = c(FALSE, "bigmemh", 5, "5G", "16:00:00", 1))
}
\arguments{
\item{inputdf}{An input data.frame, with columns of file and out. [df, cols=file, out]
Note, row parameter "j" will be passed to each array job!}

\item{outdir}{The dir of shell files. [chr, "largedata/"]}

\item{cmdno}{Number of commands to excute in each array. [num, =1]}

\item{rcodes}{The abosulte path of your R codes to run. [chr, ="lib/C_format.R"]}

\item{rversion}{R version used. [chr, ="3.5"]}

\item{arrayshid}{The sbatch id. [chr, ="slurm-script/run_bcf_query_array.sh"]}

\item{email}{Your email address that farm will email to once the jobs were done/failed. [chr, =NULL]}

\item{runinfo}{[vector, runinfo = c(FALSE, "bigmemh", 5, "5G", "16:00:00")]}
}
\value{
return a batch of shell scripts.
}
\description{
Run array rcodes according to the parameters in the inputdf file.
 The array job number = nrow(inputdf).
}
\examples{
sh <- paste0('R --no-save --args ', j, ' < ', rcodes)
run_Rcode(inputdf=data.frame(file=1:11, out=10), outdir="slurm-script", cmdno=10,
           rcodes = "lib/C_format.R", arrayshid = "slurm-script/run_rcode_array.sh",
           email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}
