% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_jobs.R
\name{set_slurm_job}
\alias{set_slurm_job}
\title{\code{Set up one farm job}}
\usage{
set_slurm_job(slurmsh = "largedata/GenSel/CL_test.sh",
  shcode = "sh largedata/myscript.sh", wd = NULL, jobid = "myjob",
  email = NULL, runinfo = c(TRUE, "bigmemh", "1", "2G", "8:00:00"))
}
\arguments{
\item{slurmsh}{Relative or absolute path and file name of your shell code, i.e. largedata/GenSel/CL_test.sh.}

\item{shcode}{The commands inside your sh file.}

\item{wd}{Working directory, default=NULL. It will use your current directory.}

\item{jobid}{The job name show up in your sq NAME column.}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{runinfo}{Parameters control the array job partition. c(TRUE, "bigmemh", "1", "2G", "8:00:00")
A vector of c(TRUE, "bigmemh", "1"): 1) run or not, 2) -p partition name, 3)  --ntasks.}
}
\value{
a shell file.
}
\description{
Farm is a computer cluster running slurm system.
Note, bigmem mem=8000/cpu, hi/med/low mem=25000/cpu, serial mem=1500/cpu.
}
\examples{
cmd <- paste0("snpconvert -a largedata/teo_updated/teo_raw_biallelic.hmp.txt",
" -i largedata/teo_updated/TeoCurated_20160308_AGPv2_flt_maf005m2.txt -s 12",
" -o largedata/teo_updated/teo_raw_biallelic_recoded_20160303_AGPv2.txt")

set_farm_job(slurmsh = "largedata/scripts/run_snpconvert.sh",
shcode = cmd, wd = NULL, jobid = "snpconvert", email=NULL, runinfo=c(TRUE, "bigmemh", "1"))

}
